customgpt-client/pyproject.toml
[tool.poetry]
name = "customgpt-client"
version = "1.1.3"
description = "A client library for accessing customgpt"

authors = []

readme = "README.md"
packages = [
    {include = "customgpt_client"},
]
include = ["CHANGELOG.md", "customgpt_client/py.typed"]

[tool.poetry.dependencies]
python = "^3.8"
sseclient-py = "1.7.2"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target_version = ['py38', 'py39', 'py310', 'py311']
exclude = '''
(
  /(
    | \.git
    | \.venv
    | \.mypy_cache
  )/
)
'''

[tool.isort]
line_length = 120
profile = "black"
